Basic JS webapp scaffold
This is a minimalist JS App. I'm going to pull in JS dependencies one at
a time as I need them in YAGNI fashion.

The CSS is using pure.css from YUI since I've been looking for a reason
to use it.

The markup is hand generated, as it's a fairly simple UI. I'm using
data-row and data-col attributes in the td's of the table to give me
coordinates for the model.